+++ render-demo/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Patch Changes
 
 - Updated dependencies
+  - @byrding/core@0.7.0
   - @byrding/react@0.3.0
 
 ## 0.0.7