+++ render-demo/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @byrding/render-demo
 
+## 0.0.9
+
+### Patch Changes
+
+- Updated dependencies [1344dab]
+- Updated dependencies [26c6a58]
+- Updated dependencies [5839498]
+- Updated dependencies [4035ce0]
+- Updated dependencies [5dca789]
+- Updated dependencies [de2aeb4]
+  - @byrding/core@0.8.0
+  - @byrding/react@0.4.0
+
 ## 0.0.8
 
 ### Patch Changes